+++ huracanpy/calc/__init__.py
@@ -12,6 +12,7 @@ __all__ = [
     "distance",
     "translation_speed",
     "azimuth",
+    "corral_radius",
 ]
 
 from ._density import density
@@ -22,4 +23,4 @@ from ._track_stats import (
 )
 from ._lifecycle import time_from_genesis, time_from_apex
 from ._rates import delta, rate
-from ._translation import distance, translation_speed, azimuth
+from ._translation import distance, translation_speed, azimuth, corral_radius